+++ lib/mobx_store_closing.dart
@@ -1,4 +1,4 @@
 library mobx_store_closing;
 
 export 'src/mobx_stores/closings.dart';
-export 'src/instantiate_stores_fake/closings.dart';
+export 'src/instantiate_stores/closings.dart';