+++ lib/mobx_store_ticket.dart
@@ -1,4 +1,4 @@
 library mobx_store_ticket;
 
 export 'src/mobx_stores/tickets.dart';
-export 'src/instantiate_stores_fake/tickets.dart';
+export 'src/instantiate_stores/tickets.dart';