+++ lib/src/mobx_stores/validators/articles/article/create_retail_form.g.dart
@@ -56,6 +56,22 @@ mixin _$ArticleRetailCreateFormStore on _ArticleCreateFormStore, Store {
     });
   }
 
+  late final _$photoPathAtom =
+      Atom(name: '_ArticleCreateFormStore.photoPath', context: context);
+
+  @override
+  String get photoPath {
+    _$photoPathAtom.reportRead();
+    return super.photoPath;
+  }
+
+  @override
+  set photoPath(String value) {
+    _$photoPathAtom.reportWrite(value, super.photoPath, () {
+      super.photoPath = value;
+    });
+  }
+
   late final _$costAtom =
       Atom(name: '_ArticleCreateFormStore.cost', context: context);