+++ test/closing/closing_ranges_test.dart
@@ -1,4 +1,5 @@
 import 'package:mixins_weebi/mobx_store_closing.dart';
+import 'package:mixins_weebi/src/instantiate_stores/closings.dart';
 import 'package:models_weebi/closings.dart' show ClosingRange;
 import 'package:test/test.dart';